+++ src/routes/MealDetailsRoute/MealDetailsRoute.js
@@ -2,10 +2,11 @@ import React, { Component } from 'react'
 import MealDetails from '../../components/MealDetails/MealDetails'
 
 class MealDetailsRoute extends Component {
+
   render() {
     return (
       <section>
-        <MealDetails />
+        <MealDetails meal_id = {this.props.match.params.meal_id}/>
       </section>
     );
   }
