+++ src/routes/MealsRoute/MealsRoute.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react'
 import Meals from '../../components/Meals/Meals'
 
 class MealsRoute extends Component {
+
   render() {
     return (
       <section>
